feat(#294): add tests and default fixture for list outputs, move list output mapping to dedicated file
